Add files via upload
We need to know HTML, CSS, and Javascript to create a responsive website. It's mandatory to learn about technologies to develop a responsive website.

First, we will start creating a static website... later will go to a responsive website. It will be easier to create a responsive website if we gain perfection in static websites.

Please follow the steps given below to create a complete of website.